requirements.txt
tensorflow-gpu==1.2.1
numpy
moviepy
requests
patool
